Prevent users from directly upgrading to the latest version
This patch to the documentation is a quick fix to prevent users from having an error when upgrading directly from 6.3.x to 7.1 (or newer) versions of Seafile image. Due to the dropped support of Python 2 in 7.1, trying to upgrade directly will fail. Changing the image tag to 7.0.5, which is the most recent version of 7.0.x branch, prevents that.

Fixes haiwen/seafile#2385